backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8
# HTTP and API clients
requests==2.31.0
httpx==0.25.2
aiohttp==3.8.5

# OAuth and Social Authentication
authlib==1.2.1
python-social-auth==0.3.6
oauthlib==3.2.2

# Data processing and ML
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1

# Environment and testing
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
boto3==1.34.0
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
Pillow==10.1.0
transformers==4.36.0
torch==2.1.1
